requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
black==22.6.0
certifi==2022.6.15
chardet==4.0.0
charset-normalizer==2.1.0
click==8.1.3
discord.ext==0.8
discord.py==1.7.3
Flask==2.1.2
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lavalink @ git+https://github.com/Devoxin/Lavalink.py.git@33195ed6246e1d33b4ee0cc79715c9fe0fb250bc
MarkupSafe==2.1.1
multidict==6.0.2
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
requests==2.28.1
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.10
Werkzeug==2.1.2
yarl==1.7.2
zipp==3.8.0
